Make new CLI commands more self-actualizing
- each command exports its own usage info
- each command parses its own cli options
- cli.js handles the --help and --version requests
