+++ cli/upgrade.js
@@ -1,12 +1,19 @@
-import { brightGreen, gray } from "../deps/colors.js";
+import {brightGreen, gray} from "../deps/colors.js";
+import {version} from "../cli.js";
 
+export const USAGE = `
+    ${brightGreen("lume upgrade")}: upgrade your local lume install to the latest
+    
+    USAGE:
+        lume upgrade
+`
 /**
  * Command to upgrade lume to the latest version
  */
-export default async function upgrade(version) {
+export default async function upgrade() {
   const response = await fetch("https://cdn.deno.land/lume/meta/versions.json");
   const versions = await response.json();
-  const { latest } = versions;
+  const {latest} = versions;
 
   if (latest === version) {
     console.log(`You're using the latest version of lume: ${latest}!`);
@@ -38,13 +45,13 @@ export default async function upgrade(version) {
 
   console.log("");
   console.log(
-    `Update successful! You're using the latest version of lume: ${
-      brightGreen(latest)
-    }!`,
+      `Update successful! You're using the latest version of lume: ${
+          brightGreen(latest)
+      }!`,
   );
   console.log(
-    "See the changes in",
-    gray(`https://github.com/lumeland/lume/blob/${latest}/CHANGELOG.md}`),
+      "See the changes in",
+      gray(`https://github.com/lumeland/lume/blob/${latest}/CHANGELOG.md`),
   );
   console.log("");
 }
